+++ client/web/src/enterprise/insights/constants/index.ts
